Refactor test suite to abstract dependencies to test-helper
