+++ package.json
@@ -7,8 +7,8 @@
     "test": "test"
   },
   "scripts": {
-    "test": "mocha -R dot",
-    "test-cov": "istanbul cover _mocha"
+    "test": "mocha 'test/**/*.js' -R dot",
+    "test-cov": "istanbul cover _mocha 'test/**/*.js'"
   },
   "repository": {
     "type": "git",
